Split dataset and model class, minor code update
